feat: Move raise hand extension out of avmoderation.
